separate loader tests
